use ava's built it file filtering (#rtfm)
